Fix breakage on big endian.
Not sure if it causes "merely" rebuilding the des cache on every startup, or
the game not working at all -- I don't have a wrong endian machine to test,
and qemu is slower than my phone.  And my phone is from 2009.

(cherry picked from commit 6a4a6370bbc2e26acc501f171d0356159a10c147)

+++ crawl-ref/source/maps.cc
@@ -45,9 +45,9 @@
 # error BYTE_ORDER is not defined
 #endif
 #if BYTE_ORDER == LITTLE_ENDIAN
-# define WORD_LEN sizeof(long)
+# define WORD_LEN (int8_t)sizeof(long)
 #else
-# define WORD_LEN -sizeof(long)
+# define WORD_LEN -(int8_t)sizeof(long)
 #endif
 
 static map_section_type _write_vault(map_def &mdef,
